LayoutTests/platform/iphone-simulator/text-autosizing/percent-adjust-number-line-height-expected.txt

Tests number line-height is adjusted with percentage -webkit-text-size-adjust values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testedLineHeight is expectedLineHeight
PASS successfullyParsed is true

TEST COMPLETE
